requirements.txt
streamlit
streamlit-autorefresh
pandas
numpy
requests
matplotlib
upstox-python-sdk
python-dotenv